+++ src/SnowRunnerTuningShop/Views/TireTuningView.xaml.cs
@@ -328,6 +328,7 @@ public partial class TireTuningView : UserControl
         private double _offRoadFriction;
         private double _mudFriction;
         private bool _ignoreIce;
+        private int _price;
 
         public required string EntryPath { get; init; }
         public required string Name { get; init; }
@@ -338,7 +339,22 @@ public partial class TireTuningView : UserControl
         public required string UsedBy { get; set; }
         public required string UsedByTooltip { get; set; }
         public required string Category { get; init; }
-        public int Price { get; init; }
+
+        public int Price
+        {
+            get => _price;
+            set
+            {
+                if (_price == value)
+                {
+                    return;
+                }
+
+                _price = value;
+                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(Price)));
+            }
+        }
+
         public required string FrictionTemplate { get; init; }
         public IReadOnlyList<TireInstanceRef> Instances { get; set; } = [];